+++ src/genetic_algorithm/fitnesses/fitness.rs
@@ -1,5 +1,5 @@
 use image::RgbaImage;
 
 pub trait Fitness {
-    fn calculate(&self, source: &RgbaImage, result: &RgbaImage) -> f32;
+    fn calculate(&self, source: &RgbaImage, result: &RgbaImage) -> f64;
 }
